refactor: Convert Programmatic UI into xib for AddTagsCell
