+++ sphinx/Scenes/Groups/Public Groups/Custom Views/AddedTagsView.xib
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="22505" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22504"/>
+        <capability name="Named colors" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <customFonts key="customFonts">
+        <array key="Roboto-Medium.ttf">
+            <string>Roboto-Medium</string>
+        </array>
+    </customFonts>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AddedTagsView" customModule="sphinx" customModuleProvider="target">
+            <connections>
+                <outlet property="container" destination="0Gb-ht-hs7" id="Ns1-FY-lcy"/>
+                <outlet property="contentView" destination="iN0-l3-epB" id="8s1-TJ-mU4"/>
+                <outlet property="tagLabel" destination="ZuL-6G-lbS" id="ZzF-wM-8wd"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="iN0-l3-epB">
+            <rect key="frame" x="0.0" y="0.0" width="300" height="50"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="0Gb-ht-hs7">
+                    <rect key="frame" x="10" y="5" width="290" height="40"/>
+                    <subviews>
+                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Bitcoin" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ZuL-6G-lbS">
+                            <rect key="frame" x="20" y="12.5" width="230" height="15.5"/>
+                            <fontDescription key="fontDescription" name="Roboto-Medium" family="Roboto" pointSize="13"/>
+                            <color key="textColor" name="Text"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="xmark" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="AEK-Ag-lL8">
+                            <rect key="frame" x="260" y="14" width="10" height="12"/>
+                            <color key="tintColor" name="SecondaryText"/>
+                            <constraints>
+                                <constraint firstAttribute="width" constant="10" id="bi9-nL-PzR"/>
+                                <constraint firstAttribute="height" constant="15" id="eeS-K5-1wP"/>
+                            </constraints>
+                            <preferredSymbolConfiguration key="preferredSymbolConfiguration" weight="bold"/>
+                        </imageView>
+                    </subviews>
+                    <color key="backgroundColor" name="Body"/>
+                    <constraints>
+                        <constraint firstItem="AEK-Ag-lL8" firstAttribute="centerY" secondItem="0Gb-ht-hs7" secondAttribute="centerY" id="ENB-xK-8xK"/>
+                        <constraint firstItem="ZuL-6G-lbS" firstAttribute="leading" secondItem="0Gb-ht-hs7" secondAttribute="leading" constant="20" id="KVD-9g-5cu"/>
+                        <constraint firstItem="ZuL-6G-lbS" firstAttribute="centerY" secondItem="0Gb-ht-hs7" secondAttribute="centerY" id="UUd-63-xSQ"/>
+                        <constraint firstAttribute="trailing" secondItem="AEK-Ag-lL8" secondAttribute="trailing" constant="20" id="XK9-QG-qjg"/>
+                        <constraint firstItem="AEK-Ag-lL8" firstAttribute="leading" secondItem="ZuL-6G-lbS" secondAttribute="trailing" constant="10" id="zzf-UV-fvK"/>
+                    </constraints>
+                </view>
+            </subviews>
+            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+            <constraints>
+                <constraint firstAttribute="trailing" secondItem="0Gb-ht-hs7" secondAttribute="trailing" id="3qb-zD-gSX"/>
+                <constraint firstItem="0Gb-ht-hs7" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="10" id="FCe-Nk-BL4"/>
+                <constraint firstItem="0Gb-ht-hs7" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="5" id="I6e-AD-wNb"/>
+                <constraint firstAttribute="bottom" secondItem="0Gb-ht-hs7" secondAttribute="bottom" constant="5" id="qNb-4t-t7e"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <point key="canvasLocation" x="137.68115942028987" y="123.88392857142857"/>
+        </view>
+    </objects>
+    <resources>
+        <image name="xmark" catalog="system" width="128" height="113"/>
+        <namedColor name="Body">
+            <color red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+        <namedColor name="SecondaryText">
+            <color red="0.55686274509803924" green="0.58823529411764708" blue="0.61176470588235299" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+        <namedColor name="Text">
+            <color red="0.23529411764705882" green="0.24705882352941178" blue="0.25490196078431371" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+    </resources>
+</document>
